domain as project
When creating a domain, create a project entry for that domain.

Many tests have different expectations with respect to the number
of projects expected.  Trying to document them all is madness, but
at least document the expected numbers.  LDAP will not support
a project for the domain, but SQL will, and there are tests that mix
both together.

Change-Id: I8bdff22e20d672d05a732723c9a35f87ba0f0063